Rename: 파일 성격에 맞게 폴더로 이동
