+++ src/domain/LottoConstoller.js
