+++ src/domain/Lottos.js
